# [storage] Fix and optimize file index calculation