## Summary

On compaction there're two mapping involved:
- old record location -> new record location, used for index remapping
- new record location -> file index, used to decide `seg_idx` inside of compacted file index

I used two hashmaps in the current implementation, with each of them storing every valid row in the map.
But the second mapping could be calculated easily via simple math, which grows sequentially.
In theory, this PR fix should cut down all hash-related operation latency into half.

Benchmark, same setup as https://github.com/Mooncake-Labs/moonlink/pull/1174
- before: 7 seconds, including 3 sec for data file, 4 sec for file indices
- after: 4.5 seconds, including 2.7 sec for data file, 1.8 sec for file indices

## Checklist

- [x] Code builds correctly
- [ ] Tests have been added or updated
- [ ] Documentation updated if necessary
- [x] I have reviewed my own changes