@dentiny 2025-08-01
Actually all tests have been passed
```sh
running 6 tests
test table_handler::chaos_test::test_chaos_with_data_compaction has been running for over 60 seconds
test table_handler::chaos_test::test_chaos_with_data_compaction_with_chaos_injection has been running for over 60 seconds
test table_handler::chaos_test::test_chaos_with_index_merge has been running for over 60 seconds
test table_handler::chaos_test::test_chaos_with_index_merge_with_chaos_injection has been running for over 60 seconds
test table_handler::chaos_test::test_chaos_with_data_compaction ... ok
test table_handler::chaos_test::test_chaos_with_no_background_maintenance has been running for over 60 seconds
test table_handler::chaos_test::test_chaos_with_index_merge ... ok
test table_handler::chaos_test::test_chaos_with_no_background_maintenance_with_chaos_injection has been running for over 60 seconds
test table_handler::chaos_test::test_chaos_with_index_merge_with_chaos_injection ... ok
test table_handler::chaos_test::test_chaos_with_data_compaction_with_chaos_injection ... ok
test table_handler::chaos_test::test_chaos_with_no_background_maintenance ... ok
test table_handler::chaos_test::test_chaos_with_no_background_maintenance_with_chaos_injection ... ok
```